Main BCI script
This script calls the extracted model from sigma, an EEG database and sends epochs to our model every 2 seconds. The output is stored and compared to the expectation. Ultimately, sensitivity, specificity usw. are computed and sent back to the user.
%%%%%
%%%%%
As of today, the SVM only outputs (1) values (i.e emergency). Since it mostly output (-1) values back in sigma, I suspect the bfeature_extraction() function is not working correctly.